+++ src/schema/slack/types/custom/interactivity.ts
@@ -1,0 +1,18 @@
+import SchemaTypes from "../../../schema_types.ts";
+import { DefineType } from "../../../../types/mod.ts";
+import UserContextType from "./user_context.ts";
+
+const InteractivityType = DefineType({
+  name: "User Context",
+  type: SchemaTypes.object,
+  properties: {
+    interactivity_pointer: {
+      type: SchemaTypes.string,
+    },
+    interactor: {
+      type: UserContextType,
+    },
+  },
+});
+
+export default InteractivityType;
