+++ src/schema/slack/types/custom/mod.ts
@@ -1,0 +1,9 @@
+import InteractivityType from "./interactivity.ts";
+import UserContextType from "./user_context.ts";
+
+const CustomSlackTypes = {
+  interactivity: InteractivityType,
+  user_context: UserContextType,
+};
+
+export default CustomSlackTypes;
