+++ src/schema/slack/types/mod.ts
@@ -1,0 +1,10 @@
+const SlackTypes = {
+  user_id: "slack#/types/user_id",
+  channel_id: "slack#/types/channel_id",
+  usergroup_id: "slack#/types/usergroup_id",
+  timestamp: "slack#/types/timestamp",
+  blocks: "slack#/types/blocks",
+  oauth2: "slack#/types/credential/oauth2",
+} as const;
+
+export default SlackTypes;
